iCalendar event export
Includes individual event exports as well exporting all the events shown
in the list at once (with the "Add All" button).

We get a bunch of package updates with this too. The `icalendar` library
wanted a newer version of `timex`, which necessitated a newer version of
`coherence`, which changes a bunch of things. Dropped dependency on
`phoenix_timex`.

Also misc. whitespace cleanup.

Resolves #22